@@ -1,3 +1,5 @@
+import type {SidebarsConfig} from '@docusaurus/plugin-content-docs';
+
 /**
  * Creating a sidebar enables you to:
  - create an ordered group of docs
@@ -8,22 +10,14 @@
 
  Create as many sidebars as you want.
  */
-
-// @ts-check
-
-/** @type {import('@docusaurus/plugin-content-docs').SidebarsConfig} */
-const sidebars = {
+const sidebars: SidebarsConfig = {
   // By default, Docusaurus generates a sidebar from the docs folder structure
-  tutorialSidebar: [
-    {type: 'autogenerated', dirName: 'tutorial'},
-  ],
-
-  apiSidebar: [{type: 'autogenerated', dirName: 'api'}],
+  guideSidebar: [{type: 'autogenerated', dirName: '.'}],
 
   // But you can create a sidebar manually
   /*
   tutorialSidebar: [
-    'introduction',
+    'intro',
     'hello',
     {
       type: 'category',
@@ -34,4 +28,4 @@ const sidebars = {
    */
 };
 
-module.exports = sidebars;
+export default sidebars;
